Sedikit perbaikan penulisan

+++ src/api/components/movies/movie-service.js
@@ -1,28 +1,44 @@
 const movieRepository = require('./movie-repository');
-const findOrFail = require('../../../utils/findOrFail');
 
-class MovieService {
-  async getAllMovies(filter) {
-    return await movieRepository.findAll(filter);
-  }
-
-  async getMovieById(id) {
-    return await findOrFail(movieRepository, id, 'Movie');
-  }
-
-  async createMovie(data) {
-    return await movieRepository.create(data);
-  }
-
-  async updateMovie(id, data) {
-    await findOrFail(movieRepository, id, 'Movie');
-    return await movieRepository.update(id, data);
-  }
-
-  async deleteMovie(id) {
-    await findOrFail(movieRepository, id, 'Movie');
-    await movieRepository.delete(id);
-  }
+async function getAllMovies(filter) {
+  return movieRepository.findAll(filter);
 }
 
-module.exports = new MovieService();
+async function getMovieById(id) {
+  const movie = await movieRepository.findById(id);
+  if (!movie) {
+    return null;
+  }
+
+  return movie;
+}
+
+async function createMovie(data) {
+  return movieRepository.create(data);
+}
+
+async function updateMovie(id, data) {
+  const movie = await movieRepository.findById(id);
+  if (!movie) {
+    return null;
+  }
+
+  return movieRepository.update(id, data);
+}
+
+async function deleteMovie(id) {
+  const movie = await movieRepository.findById(id);
+  if (!movie) {
+    return null;
+  }
+
+  await movieRepository.delete(id);
+}
+
+module.exports = {
+  getAllMovies,
+  getMovieById,
+  createMovie,
+  updateMovie,
+  deleteMovie,
+};
